Integrated Communities Graph into index.py
